made executeable. Everything for Cyberstorm deployment is in the new Test folder. need to finish doc then good to turn in
