+++ Project/Room Adventure/Demo/Test/README.txt
@@ -1,0 +1,9 @@
+Some helpful hints/tips:
+- Install the Pillow library to use thing.py with pip
+- you can LOOK at items listed that are in a room
+- Some items you can TAKE
+- And you can USE some items
+- Items in your inventory cannot be looked at but they can be used
+- You can ENTER a code in any rooms that have a visible keypad or where you are told there is a hidden keypad
+- Each keypad is programmed to accept only 1 code
+- Some codes may need to be encoded rather than decoded
